Continue refactoring vimrc
Moved most mappings into a separate file, as well as unnecessary
filetype autocmds into ftplugins. There's still a lot we can do...

One nice thing is we've added the .vim/init directory to the path
for .vimrc files, so we can `gf` on the Source lines to pop into
the relevant file.