Strip out Unite in favor of FZF
Cleans up a lot of cruft in .vimrc (though it could still stand
to be tidied a bit more) and makes file navigation quite a bit
snappier, now that MacVim finally plays nicely with FZF in the
terminal.

This is a big change, so this commit will live in a branch at first.
I've lived with Unite and all that old project browsing code for
a long time, so I want to make sure I'm happy with how all this works
before merging it into master... but I'm pretty sure I won't turn back.