Use FZF for ctrl-r history input history searching
In particular, we can move around the input line to tweak the query,
which the basic history-incremental-search-backward doesn't allow.

This widget is tweaked from the FZF default so we can use tab to "paste"
the selected item into the buffer, and enter to accept, similar to
what we're used to.