Disable jedi and old jsx plugins
Jedi seems to be confusing YCM for python files, but YCM covers its
use cases now anyway, so I could probably just remove it.

The JSX plugins seem to be interfering with TSX for some reason.
Not sure yet why it doesn't work here when it was fine on my old
laptop, but I don't really do plain JSX anyway so...