nvim: Begin working to support a "smart" K in lua scripts
For vim.* stuff (and anything in a vim.cmd() code block) we *probably*
want to open the vim :help; for lua stuff, we still want access to the
normal LSP doc lookup.

This change uses treesitter to determine the language of the AST node
under the cursor; if it's vim, we use the Scriptease help doc lookup.
Otherwise, we use the default LSP doc lookup. I believe in our next step
we can create some sort of query to check whether the expression our
lookup target is in is part of a vim.* expression....