nvim: Keep fold at current line open through formatting
For whatever reason, the method we were using before no longer works.
Did I imagine it?!

In any case, this is much more definitely correct. In the future,
perhaps we could scan the buffer for folds and preserve *all* of them.
That would probably be tricker due to code moving around on format, so
let's keep it simple for now.