Add a new "monorepo" mode for projects to make navigation faster
When opening a file that's part of a monorepo (as we've defined it, a
project whose directory name prefixes the name of subdirectory
containing the selected file, separated by `-` or `_`) we now use that
sub-directory (sub-project) as the root of the "project" for all
project navigation mappings, and for the local `path`. This makes it
much faster to navigate within the sub-project, which is much more
common than navigating across sub-projects.

On the rare occasion I need cross-monorepo file navigation, it can
still accessed under the <leader>m prefix