lazy: FINALLY fix a hang in insert mode, caused by compare.scopes
This was a massive pain to deduce, but at last I can type in elixir
files without everything grinding to a halt. I initially was testing
against opening the cmdline (IE: `:`) and disabling the signature
plugins seemed to help, but I guess that was a red herring. Disabling
this one scoring function has fixed both insert mode and cmdline mode
hangs.