lazy: Clean up otsukare import
We now conditionally import "all" lazy plugins from otsukare, so we
don't have to deal with the import errors in a bunch of different
places. We use a basic "existence" conditional for importing otsukare
here, so we don't get any annoying error popups if we happen to use this
config somewhere we haven't also installed otsukare (as it should be).