Fix trouble window not opening + customize keymaps a bit
This can be nice when we have a really long diagnostic. I don't care for
the window "following" me as I move windows (I don't want to leave this
thing open constantly anyway) and I also don't need to see all
diagnostics across the project, so the custom keymaps reflect that.